+++ src/main/java/com/futurenet/cotree/auth/security/oauth2/handler/CustomOAuth2SuccessHandler.java
@@ -1,7 +1,7 @@
-package com.futurenet.cotree.auth.oauth2.handler;
+package com.futurenet.cotree.auth.security.oauth2.handler;
 
 import com.futurenet.cotree.auth.constant.JwtConstants;
-import com.futurenet.cotree.auth.dto.UserPrincipal;
+import com.futurenet.cotree.auth.security.dto.UserPrincipal;
 import com.futurenet.cotree.auth.dto.request.RefreshTokenSaveRequest;
 import com.futurenet.cotree.auth.repository.RefreshTokenRepository;
 import com.futurenet.cotree.auth.util.JwtUtil;
