+++ src/main/java/com/futurenet/cotree/auth/service/exception/AuthErrorCode.java
@@ -1,4 +1,4 @@
-package com.futurenet.cotree.auth.exception;
+package com.futurenet.cotree.auth.service.exception;
 
 import com.futurenet.cotree.global.exception.ErrorCode;
 import lombok.AllArgsConstructor;
